+++ packages/inventory/src/Stocktake/DetailView/modal/StocktakeLineEdit/hooks/useDraftStocktakeLines.ts
@@ -1,13 +1,13 @@
 import { useState, useEffect, Dispatch, SetStateAction } from 'react';
 import { useStockLines, ItemRowFragment } from '@openmsupply-client/system';
-import { useStocktakeLines, useStocktakeFields } from './../../../../api';
+import { useStocktake } from './../../../../api';
 import { DraftStocktakeLine, get } from '../utils';
 
 export const useDraftStocktakeLines = (
   item: ItemRowFragment | null
 ): [DraftStocktakeLine[], Dispatch<SetStateAction<DraftStocktakeLine[]>>] => {
-  const { id } = useStocktakeFields('id');
-  const { data: stocktakeLines } = useStocktakeLines(item?.id);
+  const { id } = useStocktake.document.fields('id');
+  const { data: stocktakeLines } = useStocktake.line.stocktakeLines(item?.id);
   const { data: stockLines } = useStockLines(item?.id || '');
 
   const [draftLines, setDraftLines] = useState<DraftStocktakeLine[]>([]);
