+++ packages/inventory/src/Stocktake/api/hooks/document/index.ts
@@ -1,0 +1,19 @@
+import { useInsertStocktake } from './useInsertStocktake';
+import { useStocktake } from './useStocktake';
+import { useStocktakeDelete } from './useStocktakeDelete';
+import { useStocktakeDeleteSelected } from './useStocktakeDeleteSelected';
+import { useStocktakeFields } from './useStocktakeFields';
+import { useStocktakes } from './useStocktakes';
+import { useStocktakesAll } from './useStocktakesAll';
+import { useUpdateStocktake } from './useUpdateStocktake';
+
+export const Document = {
+  useInsertStocktake,
+  useStocktake,
+  useStocktakeDelete,
+  useStocktakeDeleteSelected,
+  useStocktakeFields,
+  useStocktakes,
+  useStocktakesAll,
+  useUpdateStocktake,
+};
