+++ packages/inventory/src/Stocktake/api/hooks/document/useStocktakeDelete.ts
@@ -1,0 +1,10 @@
+import { useQueryClient, useMutation } from '@openmsupply-client/common';
+import { useStocktakeApi } from '../utils/useStocktakeApi';
+
+export const useStocktakeDelete = () => {
+  const queryClient = useQueryClient();
+  const api = useStocktakeApi();
+  return useMutation(api.deleteStocktakes, {
+    onSuccess: () => queryClient.invalidateQueries(api.keys.base()),
+  });
+};
