+++ packages/inventory/src/Stocktake/api/hooks/document/useStocktakesAll.ts
@@ -1,0 +1,11 @@
+import { useMutation, SortBy } from '@openmsupply-client/common';
+import { StocktakeRowFragment } from '../../operations.generated';
+import { useStocktakeApi } from '../utils/useStocktakeApi';
+
+export const useStocktakesAll = (sortBy: SortBy<StocktakeRowFragment>) => {
+  const api = useStocktakeApi();
+
+  return {
+    ...useMutation(api.keys.sortedList(sortBy), api.get.listAll({ sortBy })),
+  };
+};
