+++ packages/inventory/src/Stocktake/api/hooks/document/useUpdateStocktake.ts
@@ -1,0 +1,10 @@
+import { useQueryClient, useMutation } from '@openmsupply-client/common';
+import { useStocktakeApi } from '../utils/useStocktakeApi';
+
+export const useUpdateStocktake = () => {
+  const queryClient = useQueryClient();
+  const api = useStocktakeApi();
+  return useMutation(api.update, {
+    onSuccess: () => queryClient.invalidateQueries(api.keys.base()),
+  });
+};
