+++ packages/inventory/src/Stocktake/api/hooks/utils/index.ts
@@ -1,0 +1,7 @@
+import { useIsStocktakeDisabled } from './useIsStocktakeDisabled';
+import { useStocktakeApi } from './useStocktakeApi';
+
+export const Utils = {
+  useIsStocktakeDisabled,
+  useStocktakeApi,
+};
