+++ packages/inventory/src/Stocktake/api/hooks/utils/useIsStocktakeDisabled.ts
@@ -1,0 +1,8 @@
+import { useStocktake } from '../document/useStocktake';
+import { isStocktakeDisabled } from '../../../../utils';
+
+export const useIsStocktakeDisabled = (): boolean => {
+  const { data } = useStocktake();
+  if (!data) return true;
+  return isStocktakeDisabled(data);
+};
